move block init to separate file
